Feature/buy confirm(#9) - 구매 확정

+++ src/main/java/com/ddang/usedauction/member/domain/Member.java
@@ -40,7 +40,7 @@ public class Member extends BaseTimeEntity {
     private boolean siteAlarm;
 
     @Column
-    private int point;
+    private long point;
 
     @Column
     private String social;
